Adds controls for defining custom faceting behavior for numerical features.
Consumers can now configure the `contextName` (the name of the module in which
this control is used), the `choiceLimit` (the number of features that can be
selected concurrently), and the `binLimit` (the number of bins created from all
feature permutations).

The `@facets-change` action sends the currently selected `features` and
numerical `bins` to the consumer. You can find examples of use in the Slice,
Metrics, and Threshold modules.

PiperOrigin-RevId: 424334109