Simplify table sorting.
The use of sticky sorting in a reaction was causing issues with template strings being correctly rendered due to interaction between observables and lit html template strings. The end result was that trying to sort tables with template strings as cells would cause javascript errors and incorrect table contents.

Since we have no client need for sticky sort, removing it to simplify the code and fix the sorting issue.

PiperOrigin-RevId: 424852922