Add DataService and use to include TCAV cosine similarity scores in scalars and data table module.
DataService stores new (non-input, non-model-output) columns of data for each datapoint in the dataset, and fetches the values for these columns when new datapoints are created.

TCAV module adds a new column when a statistically-significant CAV is created, containing the cosine similarity of each datapoint to the CAV. The TCAV interpreter is updated to be able to generate these scores from a list of datapoints and a CAV, if a pre-computed CAV is supplied to the interpreter.

The scalars module plots this new value (and any other scalars in the data service).
The data table module will this new value (and any other cols added by data service that are tagged as to be visualized in the data table).

The group service and color service also make use of the data service to take into account these new sources of information per datapoint, so datapoints can be colored and faceted by these new columns now.

PiperOrigin-RevId: 425972500