Adds the feature attribution module
This module displays a table summarizing the minimum, median, mean, and maximum
salience values for each feature in the dataset. Only `FeatureSalience` is
supported, and salience can be provided by the model (i.e., intrinsic) or
retrieved from an interpreter (e.g., derived).

Controls are provided to enable/disable the compatible salience providers, and
to facet the current dataset into subsets.

A table is created for each salience provider and each facet. Tables are shown
in an expansion panel so users can choose to show/hide data as they see fit.

PiperOrigin-RevId: 426202158