Adds faceting support to Confusion Matrix Module.
Confusion Matrix module now displays two sections of content. The first section
will always contain a confusion matrix for the entire dataset, and possibly a
matrix for the current selection, if the user enables it.

The second section displays a confusion matrix for each of the user-defined
facet groups. You can configure the faceting behavior using the faceting
control in the toolbar. Facet group matrices are sorted into ascending
alphabetical order by title, and wrap to the next row when they overflow their
parent container.

This change removes the ability to "add" new matrices manually, to simplify
the user experience. Now, all displayed matrices will have the same rows and
columns, which you can select using the drop-downs in the toolbar.

PiperOrigin-RevId: 430255969