Type-safe layout declarations in Python.
Now that enums are supported by pytype
(https://github.com/google/pytype/issues/788), we can use this to validate
layouts at build time. Use an enum to define a list of all available frontend
modules using the same names as in TypeScript, and map them to the element
names necessary for dynamic construction.

PiperOrigin-RevId: 444335966