Add TSV loader dataset.
This is one of multiple commits bringing the input salience demo to life. With this demo we're showing the results of the paper '"Will You Find These Shortcuts?" A Protocol for Evaluating the Faithfulness of Input Salience Methods for Text Classification' [https://arxiv.org/abs/2111.07367].

In this commit we introduce a class that loads data from a 2 column TSV file.

PiperOrigin-RevId: 446463349