Add a requiredForTab option in the layout spec, and basic support for per-module config.
If requiredForTab is true and that module is not compatible with the current
model(s)/data, the entire tab will be hidden. This is useful to avoid redundant
tabs, e.g. if there is a separate "Salience" tab that includes model preds, but
the model does not support any salience methods we will hide this tab.

Also preliminary support for customizing modules in the layout, using a ModuleConfig object that we can add params to, and syntactic sugar to make
  this play nice with existing config defs.

PiperOrigin-RevId: 447741156