Use more sensible default values for settings in the clustering module.
Looking at 100 top tokens isn't helpful in understanding what a cluster resembles. Therefore, we're allowing only a smaller number.

Since the goal of clustering is to find patterns in the dataset we will likely want to dissect the dataset into smaller rather than bigger portions. Because, we anticipate small subsets of data having a bias/unwanted artifact. Therefore, we're increasing the number of maximum clusters.

PiperOrigin-RevId: 459280564