Allow launching with new datapoints through POST operation.
New post_new_data endpoint to add new data for use in LIT which returns URL param to use the posted data when launching LIT and a new endpoint to remove it from the server.
Adds new URL param to load this posted data through use of the new load_new_data endpoint.

PiperOrigin-RevId: 481739685