Adds dataset and model init spec params to LitApp(...)
Refactors prior change to move all init spec info to the top level of LitMetadata. This change was made because the addition of the LitApp(...) params makes the prior assumption that there would be a 1:1 correlation between named datasets/models and an init spec impossible to guarantee. Also, this has the added benefit of making init specs much easier to find and debug in the LitMetadata JSON structure :)

Demos will be updated to make use of these params in a subsequent revision.

PiperOrigin-RevId: 506286808