Fix Datapoint Editor resize observer handling.
Previously, this could error if the Datapoint Editor was initialized but not
rendered, such as if the LIT UI loads with this module on an inactive tab. This
would cause issues with other reactions in the module.

Fix by calling .observe() inside updated() and only if the module container
element is available.

PiperOrigin-RevId: 513924936