Improve logic for header content and tooltips in the table element.
If ColumnHeader.html is nullish, generates only the HTMLDivElement necessary to display ColumnHeader.name inside the <th> element and assigns it to ColumnHeader.html.

Adds a ColumnHeader.tooltip?: string property containing the content for the tooltip. If present, ColumnHeader.htmlU will be wrapped in a LitTooltip. If nullish, then the content will only be wrapped if the content it exceeds the maximum width of the <th> element.

PiperOrigin-RevId: 521794328