Improved logic for single-line text overflow in table headers.
For instances when the DataTable element constructs the HTML for a column header, use the combination of the local (i.e., from the ColumnHeader) and global (i.e., from the table element) max widths to determine which width to use, and set that for this column using CSS custom properties and the style= attribute.

Adds a utility function for computing the width of a string (on one line) using an invisible and ephemeral canvas element to support this computation.

Converts all width specification options for table headers and tooltips to be numbers in CSS pixel units.

PiperOrigin-RevId: 521904497