Fix alignment of class selector dropdowns and search boxes in data table filtering
Having the right-aligned-search CSS rule as is causes issues with both filter text boxes and dropdowns rendering all the way to the right of the column in cases where the column width is much bigger than the header text. Changing this rule to be dependent on header.rightAlign vs. just whether it's the last column in the table fixes the alignment issues.

PiperOrigin-RevId: 529491925