Update CachingModelWrapper.predict() to use the cache instead of bypassing.
Simplifies CachingModelWrapper by removing the dataset_name and using the model name in the CacheKey.

Updates CachingModelWrapper.predict_with_metadata() to act as a passthrough to predict(), now that predict() uses the cache.

This will simplify and stabilize LIT's behavior as we move toward the removal of the Model.predict_with_metadata() API from models.

PiperOrigin-RevId: 549716363