Allow the empty dataset to be compatible with the models loaded after its initialization.
Currently, the empty dataset is initialized with all the models available at server initialization. However, if we try to load in new models after the empty dataset is initialized, they might be incompatible. Thus, we update the spec of the empty dataset every time we load in a new model.

PiperOrigin-RevId: 550675046