Remove unprotected access of AppState.currentModelSpecs.
Consumers should prefer AppState.getModelSpec() for direct access to model Specs, as it will throw an Error if it cannot find a Spec for the requested model.

Consumers may continue to reference AppState.currentModelSpecs for indirect access use cases, e.g., for reactions that need to fire after the Spec info for the active models is computed.

PiperOrigin-RevId: 559524168